Shader stage flags already aligned

+++ src/hal/src/pso/mod.rs
@@ -97,7 +97,7 @@ bitflags!(
 bitflags!(
     /// Combination of different shader pipeline stages.
     #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-    pub struct ShaderStageFlags: u16 {
+    pub struct ShaderStageFlags: u32 {
         /// Vertex shader stage.
         const VERTEX   = 0x1;
         /// Hull (tessellation) shader stage.
